add unary to expr_to_value
